requirements.txt
pandas
scikit-learn
scikit-surprise
streamlit
gradio
numpy 
joblib 
matplotlib
seaborn